+++ demos/kitchen_sink/main.py
@@ -41,7 +41,6 @@ from kivy.properties import (
     ListProperty,
     BooleanProperty,
     NumericProperty,
-    DictProperty,
 )
 from kivy.uix.modalview import ModalView
 from kivy.utils import get_hex_from_color
@@ -59,10 +58,10 @@ from kivymd.uix.list import (
     ThreeLineRightIconListItem,
 )
 from kivymd.uix.selectioncontrol import MDCheckbox
+from kivymd.theming import ThemableBehavior
 from kivymd.uix.card import MDCard
 from kivymd.utils.cropimage import crop_image
 from kivymd.utils import asynckivy
-from kivymd.theming import ThemeManager
 from kivymd.uix.dropdownitem import MDDropDownItem
 from kivymd.icon_definitions import md_icons
 
@@ -323,6 +322,7 @@ class KitchenSink(MDApp, Screens):
     create_stack_floating_buttons = BooleanProperty(False)
     manager_open = BooleanProperty(False)
     cards_created = BooleanProperty(False)
+    toolbar_hide = BooleanProperty(False)
 
     _interval = NumericProperty(0)
     tick = NumericProperty(0)
@@ -393,12 +393,16 @@ class KitchenSink(MDApp, Screens):
             f"and contributors..."
         )
         self.list_name_icons = list(md_icons.keys())[0:15]
-        Window.bind(on_keyboard=self.events)
+        Window.bind(on_keyboard=self.events, on_resize=self.window_resize)
         crop_image(
             (Window.width, int(dp(Window.height * 35 // 100))),
             f"{os.environ['KITCHEN_SINK_ASSETS']}guitar-1139397_1280.png",
             f"{os.environ['KITCHEN_SINK_ASSETS']}guitar-1139397_1280_crop.png",
         )
+
+    def window_resize(self, instance_window, width_window, height_window):
+        if self.toolbar_hide:
+            self.root.ids.toolbar.height = 0
 
     def build(self):
         self.root = Builder.load_string(root_kv)
@@ -1284,6 +1288,20 @@ class ScreenTree(BaseFanScreen):
 
 class ScreenFour(BaseFanScreen):
     pass
+
+
+class ItemBackdropBackLayer(ThemableBehavior, BoxLayout):
+    icon = StringProperty("android")
+    text = StringProperty()
+    selected_item = BooleanProperty(False)
+
+    def on_touch_down(self, touch):
+        if self.collide_point(touch.x, touch.y):
+            for item in self.parent.children:
+                if item.selected_item:
+                    item.selected_item = False
+            self.selected_item = True
+        return super().on_touch_down(touch)
 
 
 class BoxContentForBottomSheetCustomScreenList(BoxLayout):