+++ CHANGELOG.md
@@ -5,8 +5,11 @@ Change Log
 
 > [v0.102.1](https://github.com/HeaTTheatR/KivyMD/tree/0.102.1) - *Alpha*
 
+* Implemented the ability [Backdrop](https://material.io/components/backdrop)
 * Added `MDApp` class. Now app object should be inherited from `kivymd.app.MDApp`.
 * Added `MDRoundImageButton` class.
+* Added `MDTooltip` class.
+* Added `MDBanner` class.
 * Added hook for `PyInstaller` (add `hookspath=[kivymd.hooks_path]`).
 * Added examples of `spec` files for building [Kitchen Sink demo](https://github.com/HeaTTheatR/KivyMD/tree/master/demos/kitchen_sink).
 * Added some features to `MDProgressLoader`.